Side-effects are 'contagious'
Although combinations between literals and expressions with
side-effects (e.g. "something: " + foo()) technically don't
require keeping the literal, expressions with variables do
(a && foo()) and expressions with operations on the literal
do, too ("foo".myStringPrototypeExpando && foo()) and so
it seems simplest to just keep the entire thing, ugly
though it is.